reactor test
